+++ lib/lightning_web/channels/run_channel.ex
@@ -224,6 +224,18 @@ defmodule LightningWeb.RunChannel do
     end
   end
 
+  def handle_in("run:batch_logs", %{"logs" => payload}, socket) do
+    %{run: run, scrubber: scrubber} = socket.assigns
+
+    case Runs.append_run_logs_batch(run, payload, scrubber) do
+      {:error, changeset} ->
+        reply_with(socket, {:error, changeset})
+
+      {:ok, _} ->
+        reply_with(socket, :ok)
+    end
+  end
+
   # Browser client handlers
   def handle_in("fetch:run", _payload, socket) do
     run_id = socket.assigns.run_id